requirements.txt
fastapi
uvicorn
psycopg2
pydantic[email]
python-multipart
py-bcrypt
pyjwt
